Move -Wl,--gc-section and -ffast-math flags into each target
Those flags are not actually common. The -Wl,--gc-section does not exist
on Apple's llvm build but should be replaced by -Wl,-dead_strip by this
(stackoverflow post)[https://stackoverflow.com/a/24799865].
-ffast-math flag caused some numerical errors probably due to aggressive
optimizations so I disabled it for the Generic target since that one is
supposed to be more for debugging.