Fix aliasing (#125)
This PR gives a quick, not-super-optimal, but working, solution to the problem by implementing the aliasing information basically as a doubly-linked list. On a reshape, both input and output record each others name, and to check whether an alias is alive, we just traverse the aliasing double-linked list in all the directions with a bfs.

Imo for our current cases this is good enough performance-wise since there usually aren't so many aliasing chains happening, so I don't want to invest more time in the performance aspect.

As a bonus, I sprinkled a few refactors in there, the removal of bitrotten `fromVariableBuffer`s, refactor of `hoistConstant` to not use `fromVariableBuffer` (be careful with those shapes!), and refactor of TransientBuffer's `__init__` function to use its super's one (reuse and all that).

## Changed
- Removed `fromVariableBuffer`
- Refactored `hoistConstant`
- Refactored TransientBuffer's `__init__`

## Fixed
- Fixed aliasing