Remove memory-aware node bindings (#123)
Remove memory-aware node bindings because it makes the great parser refactor easier.

The memory-aware node bindings exist only for Neureka to be able to separate bindings that use the dedicated weight memory vs. those who don't.
But those bindings can simply be rewritten to check whether the weights reside in weight memory and change behavior accordingly.

By removing the memory-aware node bindings, we remove another dependency on having hoisted buffers in the middle of parsing.

The RequantHelpers are a bonus that fixes the requantization mul and add hyperrectangles to keep the rank of the original tensors.

## Added
- RequantHelpers.py for Neureka's TileConstraints

## Changed
- Removed NodeMemoryLevelChecker, MemoryAwareNodeBinding
- _parseNode from MemoryNetworkDeployer since we don't need the annotations before typeChecking anymore
- Wmem variants of bindings and tile constraints from Neureka

## Fixed
- Keep mul/add rank of requantized Neureka tile constraints